# [libc++][format][6/7] Optimizes formatted_size.